backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.7.1
torchvision==0.22.1
open_clip_torch
pillow
opencv-python
ultralytics
requests
numpy
python-dotenv
